+++ src/Sulu/Bundle/ContentBundle/Resources/public/js/components/content/form/main.js
@@ -56,6 +56,9 @@ define(['app-config', 'config', 'sulucontent/components/content/preview/main'], 
             this.sandbox.on('sulu.toolbar.save', function(action) {
                 this.submit(action);
             }, this);
+
+            // navigate away
+            this.sandbox.on('sulu.content.navigate', this.navigate, this);
         },
 
         initializeResourceLocator: function() {
@@ -477,6 +480,27 @@ define(['app-config', 'config', 'sulucontent/components/content/preview/main'], 
                     }
                 }
             ]);
+        },
+
+        navigate: function(route) {
+            var doNavigate = function(route) {
+                this.sandbox.emit('sulu.router.navigate', route);
+            }.bind(this);
+
+            if (!this.saved) {
+                this.sandbox.emit('sulu.overlay.show-warning',
+                    'sulu.overlay.be-careful',
+                    'content.template.dialog.content',
+                    function() {
+                    },
+                    function() {
+                        // ok callback
+                        doNavigate.call(this, route);
+                    }.bind(this)
+                );
+            } else {
+                doNavigate.call(this, route);
+            }
         }
     };
 });