+++ src/Sulu/Bundle/ContentBundle/Resources/public/js/components/internal-links/main.js
@@ -28,6 +28,7 @@ define([], function() {
             dataAttribute: 'internal-links',
             actionIcon: 'fa-link',
             dataDefault: [],
+            navigateEvent: 'sulu.router.navigate',
             translations: {
                 noContentSelected: 'internal-links.nolinks-selected',
                 addLinks: 'internal-links.add',
@@ -43,8 +44,12 @@ define([], function() {
                 ].join('');
             },
 
-            contentItem: function(value) {
-                return ['<span class="value">', value, '</span>'].join('');
+            contentItem: function(id, value) {
+                return [
+                    '<a href="#" data-id="', id, '" class="link">',
+                    '    <span class="value">', value, '</span>',
+                    '</a>'
+                ].join('');
             }
         },
 
@@ -59,9 +64,23 @@ define([], function() {
          * custom event handling
          */
         bindCustomEvents = function() {
-            this.sandbox.on('husky.overlay.internal-links.' + this.options.instanceName + '.add.initialized', initColumnNavigation.bind(this));
+            this.sandbox.on(
+                'husky.overlay.internal-links.' + this.options.instanceName + '.add.initialized',
+                initColumnNavigation.bind(this)
+            );
 
             this.sandbox.on('husky.column-navigation.' + this.options.instanceName + '.action', selectLink.bind(this));
+
+            this.sandbox.dom.on(this.$el, 'click', function(e) {
+                var id = this.sandbox.dom.data(e.currentTarget, 'id');
+
+                this.sandbox.emit(
+                    this.options.navigateEvent,
+                    'content/contents/' + this.options.webspace + '/' + this.options.locale + '/edit:' + id + '/details'
+                );
+
+                return false;
+            }.bind(this), 'a.link');
         },
 
         /**
@@ -198,7 +217,10 @@ define([], function() {
         },
 
         getItemContent: function(item) {
-            return templates.contentItem(item.title);
+            return templates.contentItem(
+                item[this.options.idKey],
+                item.title
+            );
         },
 
         sortHandler: function(ids) {